Improve Voip pipeline stability
It appears the pipeline is being unexpectedly cancelled in some
instances. In order to mitigate this issue hang ups will be detected
using a separate task rather than relying on timeouts in the STT read
method. The pipeline will also catch and log any CancelledErrors to help
with further debugging.